Move Module export to SwiftUIModalDelegate extension
To be used later for Characters
